WSTEAM1-691: Add Amp Iframe Story. Remove redundant oEmbed Story file
