+++ src/app/components/Embeds/EmbedError/index.stories.tsx
@@ -6,7 +6,7 @@ export default {
   component: EmbedError,
 };
 
-export const RiddleErrorDefault = () => (
+export const EmbedErrorMessage = () => (
   <EmbedError
     message="Sorry, we can’t display this part of the story on this lightweight mobile page."
     link={{
